Fixed several things related to reproducibility and standardization
Clearing keras session results in the graph of operations (and the
related data) being lost, so I did the following which resulted in 100%
reproducibility and standardization of models and data splitting across
the testing.

Models are now initialised with same weights
- previously the last layer was initialised with random weights without
a seed

Splitting of dataframes gives the same values every time
- Previously, the data would be split differently for the same user if
run n number of times

The fit method used to shuffle the training data before fitting
- this has been disabled now

These changes result in reprodicibility. Initialising users and running
train_fed with the same strategies now results in identical final
values.