Several changes working towards p2p implementation, see details
Changes to the notebook made to reflect the changes made to the .py
files

Changes to user.py
* train method now can take weights as a dict or list, if list then it
will replace the user's model with the weights provided (central strat
uses this), if a dict is supplied then the user's averaging method is
called and the weights dict is passed through (the dict is
users_id:weights)
* attributes, getters and setters for averaging_method_accuracy,
averaging_method_loss and averaging_method (this is a function pointer
to an Average class function). The accuracy/loss attributes are subject
to change.

Changes to average.py
* latest_user_metric was changed from a lambda to a private method
called _latest_user_metric
* other methods were renamed to TYPE_central or TYPE_personalised (for
p2p strat). A method called TYPE will be used as a wrapper for it, as is
done for the 'all' type averaging.
* the signature of the TYPE_STRAT methods have changed to include user
and weights(this is a dict). This may change