p2p version of std_dev implemented, pending tests Changes to user.py * in evaluate, removed the option to pass in a model. Only evaluates users model on users test data
Changes to average.py
* added evaluate_on_all_users which returns a set of user_id:metric
where metric is accuracy or loss based on what boolean parameters were
passed in, both are not allowed. it places weights from 'weights' in the
users current model and evaluates the custom weight
model on the users own test data and returns the metric as a dict.
* Added std_dev which wraps the central and personalised methods.
Central works on the metrics recieved from each user for evaluating
their models on their own test data, whereas the personalised strat
works on the user evaluating their test data on everyone elses model.
This is the p2p strat essentially. The logic for both is same, discard
the users weights if the evaluation is 1 std_dev less than the average
of the selected metric