Fixed weights not updating issue and minor changes
Fixed weights that were not updating leading to extraordinarily high accuracy,
in the file src/refactored_code_2.ipynb

Split up code to have an inits.py module for reusability and cleaner
notebooks.

Split_data_frame method changed so that the val and test percentages are
a percentage of the whole dataset and not intermediary resulting data.
Better for understanding.