Images dataset works
Results should be reproducible as seeds are being used.

init_conv_model function in "inits.py" now has a convolutional model
that works relatively well (runs faster on GPU compared to CPU).

If you run the whole "images.ipynb" notebook, then it runs out
of memory. This is why I reset the variables with regex "^users_*"
to get rid of the user models from the memory.