Updated gestures dataset output
For the gestures dataset, fixed the model that was wrong because of the dummy first record in the
dataset. Removed the dummy record as well.

The initialisation of users had a typo in the check which was fixed as
well. It works now as expected.

Minor changes to the images notebook, no reruns needed. Changed the
epochs to 16 from 1.

The models now have an input shape, required for model summary.

notebooks now have a way to get the model summary as well.